Added some British formats to the class
The formats added are needed for a current project. I have also reformatted the comments so the example is first and the description comes at the end after a comma.

Two other queries about this process... is there any appetite to create a more robust mapper between ICU and Moment?
http://userguide.icu-project.org/formatparse/datetime#TOC-Date-Time-Format-Syntax
http://momentjs.com/docs/#/displaying/format/

And if the class name in date.xml could be parameterised then it would be possible (easier) to use a local variant of MomentFormatConverter instead of editing the bundle. Then again, perhaps this way encourages more community action ;)