prepend config no matter how things are organised
Mea culpa : I did not understand why there were several config blocks
when writing this piece of code, but now it is clear : in my
application, I am the configuring sonata admin security part in
security.yml, and the other parts in config.yml . The part in security
is the first config block, but it does not contain the options key (the
part in config.yml does). This change will prepend the config as soon as
the requested config key is detected, no matter whether the
configuration is split or not. If this key appears twice in different
config blocks, the last occurrence will have its value taken into
account.