avoid converting escaped character sequences
Some character sequences that are not meant to be ignored by the date formatter
are escaped with quotes. The intl to moment converter was too naive before and
converted them. This change avoids this conversion, and escapes these character
sequences with brackets, as required by the format used by moment.js