Trigger autoload for deprecated types
By loading the deprecated types, we make sure the class alias they
contain exist, so that code type hinting against the old type accepts
the new. This is necessary because type hinting does not trigger
autoload, since it usually does not need to.
Also, we conditionally wrap the deprecation in an if statement checking
whether the new type already is loaded. That way we get the deprecation
only when really necessary.